requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.7
anthropic==0.18.1
python-dotenv==1.0.1
pydantic==2.6.1
google-generativeai==0.4.0
pyjwt==2.8.0